Moved ResourceHolder into client module
